+++ algoritms.h
@@ -16,10 +16,10 @@ void errormsg();
 figure** InitBoard(int n);
 int Safe(figure** board, figure fig, int row, int col, int n);
 figure AddFigure(char* filename);
-long long solveAll(FILE* f, figure** board, figure fig, int n, int placed, int startCell);
+uint64_t solveAll(FILE* f, figure** board, figure fig, int n, int placed, int startCell);
 void FreeBoard(figure** board, int n);
 int isAttacking(figure fig, int row1, int col1, int row2, int col2);
 void solveLocalConflicts(FILE* f, figure fig, int n, int maxSteps);
-long long solveRowByRow(FILE* f, figure** board, figure fig, int n, int row);
+uint64_t solveRowByRow(FILE* f, figure** board, figure fig, int n, int row);
 
 #endif
